Permit scanning over NAT
In order to receive results from scan targets over an L4/L3 network
address translation, the JNDI URL string must contain the DNAT IP
address and the port being mapped back to the module's LDAP server.

To make this work, the attacker must set up a port-forward on an
address routable from the victim machine and the JNDI string must
be formatted with the correct host:port pair to receive the LDAP
requests containing our scan responses.

The DNAT configuration is up to the user.
The JNDI URL can now be modified by using the added advanced config
options for DNATHOST and DNATPORT.

Testing:
  Verified that the JNDI URL was being set correctly using proper
DNATHOST and DNATPORT values and that LDAP traffic was received
on the listener with a properly configured DNAT and ACE at the edge
firewall.