Fix PSH WIP
This contains the work in progress code that is incomplete for making
the Powershell sessions over WinPTY work with Metasploit's Post API.

The issues mostly revolve around the complexity of the data sent back
from the server and identifying messages that should be filtered out
before being returned to the Post API which often inspects the results
and performs string matching. The WinPTY side would send back ANSI
control codes and would resend the prompt at odd times, particularly
when dealing with large lines (such as would be encountered while
uploading a file with the Post API).